Extract struct to models module
