+++ vit-servicing-station-cli/src/ideascale/mod.rs
@@ -1,1 +1,2 @@
 mod fetch;
+mod models;
